Organized program into multiple *.c files matching each of the function calls, also created a file called sharedFunctions.c for functions which are shared between files. Edited the makefile to compile with this. Moved some functions around. parse_file() is located in getCalendarSuite.c.
