--- core/parser.h
@@ -1,1 +1,0 @@
-int parse_file(char *);
